requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
yfinance>=0.2.36
python-dotenv>=1.0.0
apscheduler>=3.10.0
pandas>=2.0.0
